LPD-98159 Split the tree and drop-position utils by function
Each utility now lives in its own file named after the function it exports. The tree helpers move from a single tree.ts into a util/tree directory with one function per file, and every consumer imports each helper directly from its own file rather than a shared module. dropPosition.ts becomes getDropPosition.ts. The helpers that were previously private are now exported so they can stand alone.
