+++ modules/apps/audiences/audiences-web/src/main/resources/META-INF/resources/js/util/getDropPosition.ts
