+++ modules/apps/audiences/audiences-web/src/main/resources/META-INF/resources/js/util/tree/duplicateRule.ts
@@ -1,0 +1,19 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {Group} from '../../types';
+import {cloneNode} from './cloneNode';
+import {insertAt} from './insertAt';
+import {splitPath} from './splitPath';
+import {updateGroup} from './updateGroup';
+
+export function duplicateRule(root: Group, path: number[]): Group {
+	const [groupPath, index] = splitPath(path);
+
+	return updateGroup(root, groupPath, (group) => ({
+		...group,
+		items: insertAt(group.items, index + 1, cloneNode(group.items[index])),
+	}));
+}
