+++ modules/apps/audiences/audiences-web/src/main/resources/META-INF/resources/js/util/tree/isGroup.ts
@@ -1,0 +1,10 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {CriteriaNode, Group} from '../../types';
+
+export function isGroup(node: CriteriaNode): node is Group {
+	return Array.isArray((node as Group).items);
+}
